Mark IT_Exchange_Subscription constructor as private.
Introduces a static IT_Exchange_Subscription::from_transaction() method
to construct a subscription object.

The next version of Recurring Payments will change how subscriptions are
stored in the database. When this happens, we'll most likely want to change
how Subscription objects are constructed. Marking the constructor as
private will make it possible to make this change without breakage.